Upload to cloud/requirements.txt
fastapi
uvicorn
google-cloud-storage
motor
asyncpg
slowapi
python-multipart
python-dotenv
locust